+++ src/main/java/org/rorb/dev/hyphen/controllers/HyphenSuperDeveloperController.java
@@ -6,10 +6,9 @@ import org.rorb.dev.hyphen.entity.HyphenUser;
 import org.rorb.dev.hyphen.repository.DataBaseMethods;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
+import java.util.Objects;
 import java.util.Optional;
 
 @RestController
@@ -23,15 +22,33 @@ public class HyphenSuperDeveloperController {
     }
 
     @GetMapping("/org.roundrobin/hyphen/super/dev/password/38")
-    public ResponseEntity<Iterable<HyphenUser>> getAllData() {
-        Iterable<HyphenUser> message = MYSQL.findAll();
-        return ResponseEntity.of(Optional.of(message));
+    public ResponseEntity<Iterable<HyphenUser>> getAllData(@RequestHeader(value = "UserName") String userName, @RequestHeader(value = "Password") String password) {
+        if (Objects.equals(userName, "developers.hyphen") && Objects.equals(password, "prerna6938269")) {
+            Iterable<HyphenUser> message = MYSQL.findAll();
+            return ResponseEntity.of(Optional.of(message));
+        } else {
+            return ResponseEntity.status(401).build();
+        }
     }
 
-    @DeleteMapping("/org.roundrobin/hyphen.dev/super/dev/all/rorb.dev.hyphen/password/38")
-    public ResponseEntity<HyphenUser> deleteAllData() {
-        MYSQL.deleteAll();
-        return ResponseEntity.status(204).build();
+    @PostMapping("/org.roundrobin/hyphen/super/dev")
+    public ResponseEntity<HyphenUser> postChat(@RequestBody HyphenUser hyphenUser, @RequestHeader(value = "UserName") String userName, @RequestHeader(value = "Password") String password) {
+        if (Objects.equals(userName, "developers.hyphen") && Objects.equals(password, "prerna6938269")) {
+            MYSQL.save(hyphenUser);
+            return ResponseEntity.status(201).build();
+        } else {
+            return ResponseEntity.status(401).build();
+        }
+    }
+
+    @DeleteMapping("/org.roundrobin/hyphen/super/dev/all/rorb.dev.hyphen/password/38")
+    public ResponseEntity<HyphenUser> deleteAllData(@RequestHeader(value = "UserName") String userName, @RequestHeader(value = "Password") String password) {
+        if (Objects.equals(userName, "developers.hyphen") && Objects.equals(password, "prerna6938269")) {
+            MYSQL.deleteAll();
+            return ResponseEntity.status(204).build();
+        } else {
+            return ResponseEntity.status(401).build();
+        }
     }
 
 }
